Renamed "request" module to "requests" and updated references
To better adhere to the naming convention and improve code readability, the "request" module was renamed to "requests". All functions and methods that used/requested the renamed module have been updated to reflect this change. The modules in src/structure and src/function were affected. Names in other templates and issues were also amended accordingly to maintain naming consistency across the project.
