+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,5 +1,5 @@
 ---
-name: Feature request
+name: Feature requests
 about: Suggest an idea for this project
 title: ''
 labels: 'enhancement'
