+++ src/cmd/ai_module/translation.rs
@@ -75,7 +75,7 @@ pub async fn run(
             return;
         }
 
-        let allowed_extensions = vec!["mp3", "mp4", "mpeg", "mpga", "m4a", "wav", "webm"];
+        let allowed_extensions = ["mp3", "mp4", "mpeg", "mpga", "m4a", "wav", "webm"];
         let parsed_url = Url::parse(&content).expect("Failed to parse URL");
         let path_segments = parsed_url
             .path_segments()
@@ -150,7 +150,7 @@ pub async fn run(
             Ok(res) => res,
             Err(err) => {
                 let _ = fs::remove_file(&file_to_delete);
-                eprintln!("Error sending the request: {}", err);
+                eprintln!("Error sending the requests: {}", err);
                 error_making_request_edit(ctx, command, message).await;
                 return;
             }
