+++ src/cmd/mod.rs
@@ -1,5 +1,5 @@
 pub mod ai_module;
 pub mod anilist_module;
-pub mod games_module;
+//pub mod games_module;
 pub mod general_module;
-pub mod moderation_module;
+//pub mod moderation_module;
