+++ src/function/mod.rs
@@ -1,5 +1,5 @@
 pub mod anilist;
 pub mod error_management;
 pub mod general;
-pub mod request;
+pub mod requests;
 pub mod sql;
