+++ src/structure/anilist/character/struct_character.rs
@@ -1,6 +1,6 @@
 use crate::function::general::html_parser::convert_to_discord_markdown;
 use crate::function::general::trim::trim;
-use crate::function::request::request::make_request_anilist;
+use crate::function::requests::request::make_request_anilist;
 use crate::structure::embed::anilist::struct_lang_character::CharacterLocalisedText;
 use serde::Deserialize;
 use serde_json::json;
