+++ src/structure/anilist/staff/struct_staff.rs
@@ -1,6 +1,6 @@
 use crate::function::general::html_parser::convert_to_discord_markdown;
 use crate::function::general::trim::trim;
-use crate::function::request::request::make_request_anilist;
+use crate::function::requests::request::make_request_anilist;
 use crate::structure::embed::anilist::struct_lang_staff::StaffLocalisedText;
 use serde::Deserialize;
 use serde_json::json;
